Fixed build errors in VS2010, and added MSVC 2010 project.

+++ src/discimage.cpp
@@ -86,7 +86,7 @@ DiscImage::DiscImage( const char* pOutput, const char* pInput )
 		}
 
 		m_inputFile.seekg( 0, ios::end );
-		int length = m_inputFile.tellg();
+		int length = static_cast< int >( m_inputFile.tellg() );
 		m_inputFile.seekg( 0, ios::beg );
 
 		assert( length >= endSectorAddr * 0x100 );
